+++ PCIT_core/internal/llvm_interface/include/class_impls/native_ptr_decls.h
@@ -37,5 +37,6 @@ namespace llvm{
 	class Constant;
 	class ConstantInt;
 	class CallInst;
+	class LoadInst;
 	
 }