+++ PCIT_core/internal/llvm_interface/include/class_impls/stmts.h
@@ -25,17 +25,5 @@ namespace pcit::llvmint{
 		private:
 			llvm::BasicBlock* _native;
 	};
-
-
-	class Alloca{
-		public:
-			Alloca(llvm::AllocaInst* native_stmt) : _native(native_stmt) {};
-			~Alloca() = default;
-
-			EVO_NODISCARD auto native() const -> llvm::AllocaInst* { return this->_native; }
-	
-		private:
-			llvm::AllocaInst* _native;
-	};
 	
 }